Time: 17 ms (8.23%), Space: 8.7 MB (13.23%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -3,7 +3,7 @@ public:
     int mySqrt(int x) {
         int result=0;
         for(int i=1;i<=x; i++){
-            if((double)i*i<=x) {
+            if((long long)i*i<=x) {
                 result= i;
             }
             else{
